requirements.txt
Flask>=2.2,<3.0
Flask-Dance<5.0
Flask-SQLAlchemy==3.0.0
Flask-Sitemap==0.3.0
python-dotenv==1.0.0
gunicorn==23.0.0
torch==2.2.2
transformers==4.46.1
spacy==3.7.6
Werkzeug>=2.2,<2.3
google-cloud-storage==2.18.2
